agents/requirements.txt
gitpython==3.1.40
google-adk>=1.10.0
google-cloud-logging
google-cloud-storage
python-dotenv
requests
uvicorn
fastapi
flask
flask-cors
nest-asyncio>=1.6.0
httpx
google-genai==1.*  # Required for File Search API
google-generativeai  # Keep for compatibility